requirements.txt
alembic==1.8.1
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
dataclasses-json==0.5.7
docopt==0.6.2
fastapi==0.79.0
flake8==4.0.1
greenlet==2.0.2
h11==0.14.0
httplib2==0.21.0
idna==3.4
iniconfig==2.0.0
Inject==4.3.1
jmespath==1.0.1
jsons==1.6.3
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mccabe==0.6.1
msgpack==1.0.5
mypy==0.981
mypy-extensions==1.0.0
packaging==23.0
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.22.1
psutil==5.9.1
psycopg2-binary==2.9.3
py==1.11.0
py-automapper==1.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydantic==1.9.2
pyflakes==2.4.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.1.2
pytest-dotenv==0.5.2
pytest-mock==3.8.2
pytest-watch==4.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.39
starlette==0.19.1
tabulate==0.8.10
tomli==2.0.1
types-urllib3==1.26.25.8
typing-inspect==0.8.0
typing_extensions==4.5.0
typish==1.9.3
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.18.2
watchdog==2.3.1
